rename raw_logger.cc
