+++ release/files_common
@@ -308,8 +308,8 @@ selfdrive/loggerd/frame_logger.h
 selfdrive/loggerd/logger.cc
 selfdrive/loggerd/logger.h
 selfdrive/loggerd/loggerd.cc
-selfdrive/loggerd/raw_logger.cc
-selfdrive/loggerd/raw_logger.h
+selfdrive/loggerd/ffmpeg_encoder.cc
+selfdrive/loggerd/ffmpeg_encoder.h
 selfdrive/loggerd/include/msm_media_info.h
 
 selfdrive/loggerd/__init__.py
